requirements.txt
python-dotenv
proxmoxer
urllib3
requests